+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "lib/course_contract_tests"]
+	path = lib/course_contract_tests
+	url = https://github.com/umn-asr/course_contract_tests.git
